feat: options to adjust gridtk list output to fit terminal width (#15)
Fixes #13. By default, `gridtk list` outputs a table which migh not fit the terminal width.
This PR adds options to adjust the output using the `--wrap` and `--truncate` flags. The `--wrap`
flag wraps the output to fit the terminal width, while the `--truncate` flag
truncates the output to fit the terminal width.